apple-specific: Add direct assigned volume support
Today, filesystem volumes are passed in to the guest VM utilizing
virtio-fs. In a typical flow, a CSI controller will make sure backing
storage is attached to the node, and then a CSI node driver will create
the filesystem (if needed) and mount it.

Ideally, for a VM, we will consume the block device itself and not
utilize the host mount. This simplifies volume management on the host,
and is more efficient from a storage IO perspective. To do this, the CSI
node driver will need to be enlightened that the volume will be consumed
by a VM-based runtime, and will need to communicate with the runtime to
describe where the block device lives, what filesystem is on it, and
where it should be mounted within the container(s).

This mechanism will be handled within a mounted volume on the host
itself which will be empty except for a 'special' file: csiPlugin.json.

The runtime will check the container mounts for this scenario, and if
this occurs, it'll update the mount source and filesystem information
from what it reads from the json file. Based on this, the runtime will
pass in to the guest VM as a block device, and will update the storage
information consumed by the agent to mount this at an appropriate
location in the guest.

Signed-off-by: Eric Ernst <eric_ernst@apple.com>
Signed-off-by: Maksym Pavlenko <mxpv@apple.com>