refactoring continued
We are currently looping over the container mount objects, which have
lsightly more information to help with identifying a block device versus
oethers.

unfortunately ,each handler will generally also need to update the OCI
spec as well, which'll be its own loop at the end for each mount. Hard
to really avoid this unless we end up having a second global loop for
spec updates, but that seems ... kind of painful?

We at least need to special case the blk devices -- not sure how we
would differentiate these versus shared FS mounts otherwise? Are the
non-bind and shm bind types the only ones that end up looping over just spec? If so, can we handle these as a second pass?

and still have the loop for updating spec based on the mounts that come
from the host? (pass this as param for the sharedfs and block handlers?)

We still need to go through and loop through oci spec for each one
anyway....

Signed-off-by: Eric Ernst <eric_ernst@apple.com>